# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TeklifPro API",
    "version": "1.0.0",
    "description": "TeklifPro is a Turkish B2B proposal management platform designed to streamline the creation, management, and delivery of business proposals. The API provides comprehensive endpoints for authentication, proposal management, customer management, product catalogs, AI-powered suggestions, and integration with accounting systems via Parasut.",
    "contact": {
      "name": "Sercan Hosgoerns",
      "email": "hosgorens@gmail.com"
    },
    "license": {
      "name": "All Rights Reserved"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.teklifpro.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register new user and create tenant",
        "description": "Registers a new user account and creates a new tenant organization. Returns JWT token for immediate authentication. | Yeni bir kullanıcı hesabı kaydeder ve yeni bir kiracı organizasyonu oluşturur.",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "firstName",
                  "lastName",
                  "companyName"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "sercan@teklifpro.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "minLength": 8,
                    "example": "SecurePass123!"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "Sercan"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Hosgoerns"
                  },
                  "companyName": {
                    "type": "string",
                    "example": "TeklifPro Inc."
                  },
                  "phone": {
                    "type": "string",
                    "example": "+90 555 123 4567"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "user": {
                      "id": "usr_123456",
                      "email": "sercan@teklifpro.com",
                      "firstName": "Sercan",
                      "lastName": "Hosgoerns"
                    },
                    "tenant": {
                      "id": "tnt_123456",
                      "name": "TeklifPro Inc."
                    },
                    "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "refreshToken": "refresh_token_here"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Request email verification",
        "description": "Sends a verification email to the user. Called after registration to verify email address. | Kullanıcıya doğrulama e-postası gönderir.",
        "operationId": "sendVerificationEmail",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "sercan@teklifpro.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification email sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "success": true,
                  "message": "Verification email sent"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify email token",
        "description": "Verifies the email token sent in verification email. User clicks link in email which calls this endpoint. | E-posta doğrulama belirtecini doğrular.",
        "operationId": "verifyEmailToken",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "success": true,
                  "message": "Email verified successfully"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from registration or login"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "Invalid input parameters"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "usr_123456"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Tenant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "tnt_123456"
          },
          "name": {
            "type": "string"
          },
          "logo": {
            "type": "string",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/User"
              },
              "tenant": {
                "$ref": "#/components/schemas/Tenant"
              },
              "accessToken": {
                "type": "string"
              },
              "refreshToken": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}